Fixed the interference problem with Ember while changing the DOM
When we were changing the DOM (highlighting experience sentence for example)
we used to interfere with LinkedIn's front-end while it's juggling with nodes
a bajillion times per second. Concretely, and what I presumably think is our
case, Ember was trying to remove a child that we happened to alter and thus
couldn't find it.

I understood that from an exception thrown by Ember saying it coudn't call
removeChild on...